Move wait_for_command helper function to conftest.py
